requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
geoip2==5.1.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
maxminddb==2.8.2
mdurl==0.1.2
multidict==6.6.4
narwhals==2.6.0
networkx==3.5
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pefile==2024.8.26
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
rpds-py==0.27.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stix2==3.0.1
stix2-patterns==2.0.0
streamlit==1.50.0
taxii2-client==2.3.0
tenacity==9.1.2
tldextract==5.3.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
yara==1.7.7
yarl==1.20.1
